+++ src/main/java/team/skadi/powersellsys/pojo/Power.java
@@ -10,17 +10,17 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Power {
-    private Integer id;
-    private String name;
-    private Integer type;
-    private String model;
-    private Integer capacity;
-    private Integer stock;
-    private Double price;
-    private Byte discount;
-    private Integer status;
-    private LocalDateTime createTime;
-    private LocalDateTime updateTime;
-    private String describe;
-    private Integer sum;
+	private Integer id;
+	private String name;
+	private Integer type;
+	private String model;
+	private Integer capacity;
+	private Integer stock;
+	private Float price;
+	private Byte discount;
+	private Integer status;
+	private LocalDateTime createTime;
+	private LocalDateTime updateTime;
+	private String describe;
+	private Integer sum;
 }
